refactor: move utilities to domain specific directories
